+++ src/main/java/cc/spring/controllers/AdminController.java
@@ -1,24 +1,17 @@
 package cc.spring.controllers;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
-import cc.spring.dto.MemberDTO;
-import cc.spring.services.AdminMemberService;
-
-@Controller
-@RequestMapping("/adminPage/")
+@RestController
+@RequestMapping("/data/")
 public class AdminController {
 	
-	@RequestMapping("/**")
-	public String toAdmin() {
-		System.out.println("하잉");
-		return "forward:/admin/index.html";
+	@RequestMapping("business")
+	public int countBusinessMember() {
+		int countBusinessMember = 55;
+		return countBusinessMember;
 	}
+	
 	
 }
