+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -18,7 +18,7 @@
 		up static resources in the ${webappRoot}/resources directory -->
 	<resources mapping="/resources/**" location="/resources/" />
 	<!-- react -->
-	<resources mapping="/admin/**" location="/resources/adminPage/" />
+	<resources mapping="/adminPage/**" location="/resources/admin/" />
 
 	<!-- Resolves views selected for rendering by @Controllers to .jsp resources 
 		in the /WEB-INF/views directory -->
