+++ src/main/webapp/resources/admin/logo192.png
