+++ src/main/webapp/resources/admin/manifest.json
