--- src/main/webapp/resources/adminPage/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-<meta charset="utf-8" />
-<link rel="icon" href="/admin/favicon.ico" />
-<meta name="viewport" content="width=device-width,initial-scale=1" />
-<meta name="theme-color" content="#000000" />
-<meta name="description"
-	content="Web site created using create-react-app" />
-<link rel="apple-touch-icon" href="/admin/logo192.png" />
-<link rel="manifest" href="/admin/manifest.json" />
-<title>React App</title>
-<script defer="defer" src="/admin/static/js/main.287151fa.js"></script>
-<link href="/admin/static/css/main.d20202a4.css" rel="stylesheet">
-</head>
-<body>
-	<noscript>You need to enable JavaScript to run this app.</noscript>
-	<div id="root"></div>
-</body>
-</html>
